Drag and Drop: adjust offset when hovering over snippet list
I don't know where the original number came from, but in practice with
some of our smaller snippets (e.g. checkbox) this configuration made it
harder to drop it into the bottom slot.

This adjusts what range of vertical pixels the cursor needs to be at to
drop a snippet in particular spot. 30 is just a random number that felt
better than 90 when I tried it out.